Refactor LoginControllerTest imports and remove unused code
