+++ api/app/src/test/kotlin/packit/integration/controllers/LoginControllerTest.kt
@@ -4,13 +4,13 @@ import com.fasterxml.jackson.databind.ObjectMapper
 import org.assertj.core.api.Assertions.assertThat
 import org.junit.jupiter.api.Test
 import org.springframework.boot.test.web.client.TestRestTemplate
-import org.springframework.http.*
-import org.springframework.test.context.TestPropertySource
-import org.springframework.test.context.jdbc.Sql
+import org.springframework.http.HttpEntity
+import org.springframework.http.HttpHeaders
+import org.springframework.http.MediaType
+import org.springframework.http.ResponseEntity
 import packit.integration.IntegrationTest
 import packit.model.LoginWithPassword
 import packit.model.LoginWithToken
-import kotlin.test.assertEquals
 
 class LoginControllerTestGithub : IntegrationTest()
 {
@@ -52,46 +52,6 @@ class LoginControllerTestGithub : IntegrationTest()
     }
 }
 
-@TestPropertySource(properties = ["auth.method=basic"])
-@Sql("/set-test-users.sql", executionPhase = Sql.ExecutionPhase.BEFORE_TEST_METHOD)
-@Sql("/delete-test-users.sql", executionPhase = Sql.ExecutionPhase.AFTER_TEST_METHOD)
-class LoginControllerTestBasic : IntegrationTest()
-{
-    @Test
-    fun `returns token on successful basic login`()
-    {
-        val result =
-            LoginTestHelper.getBasicLoginResponse(LoginWithPassword("admin@example.com", "password"), restTemplate)
-
-        assertEquals(result.statusCode, HttpStatus.OK)
-        assertThat(result.body).contains("token")
-    }
-
-    @Test
-    fun `returns forbidden when github login is disabled`()
-    {
-        val result = LoginTestHelper.getGithubLoginResponse(LoginWithToken("token"), restTemplate)
-        assertForbidden(result)
-    }
-
-    @Test
-    fun `returns unauthorized when user user is not found`()
-    {
-        val result =
-            LoginTestHelper.getBasicLoginResponse(LoginWithPassword("notexists@example.com", "password"), restTemplate)
-
-        assertUnauthorized(result)
-    }
-
-    @Test
-    fun `returns unauthorized when password is incorrect`()
-    {
-        val result =
-            LoginTestHelper.getBasicLoginResponse(LoginWithPassword("admin@example.com", "notcorrect"), restTemplate)
-
-        assertUnauthorized(result)
-    }
-}
 
 object LoginTestHelper
 {
